Move ExcerptOverflowMonitor to src/sidebar/util
This module previously lived under `src/sidebar/directive` but it isn't
a directive, only a helper class used by a directive.
