+++ src/sidebar/util/excerpt-overflow-monitor.js
