+++ src/sidebar/util/test/excerpt-overflow-monitor-test.js
